LayoutTests/fast/js/primitive-property-access-edge-cases-expected.txt

This page tests for assertion failures in edge cases of property lookup on primitive values.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS didNotCrash is true
PASS successfullyParsed is true

TEST COMPLETE
